requirements.txt
# Audio - WASAPI
PyAudioWPatch>=0.2.12.6

# Transcription
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Speaker ID (no HuggingFace auth needed!)
resemblyzer>=0.1.3
librosa>=0.10.0
soundfile>=0.12.0

# Analysis
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
textstat>=0.7.3
scipy>=1.11.0
numpy>=1.24.0

# Claude API
anthropic>=0.18.0

# Database
sqlalchemy>=2.0.0

# UI
pystray>=0.19.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0

# Voice Activity Detection (required by resemblyzer)
webrtcvad>=2.0.10
